chores: move and rename qml parser code
